create searchBar and Error View
