--- WeatherForecastApp/WeatherForecastApp/ContentView.swift
@@ -1,24 +1,0 @@
-//
-//  ContentView.swift
-//  WeatherForecastApp
-//
-//  Created by Lawrence on 10/26/23.
-//
-
-import SwiftUI
-
-struct ContentView: View {
-    var body: some View {
-        VStack {
-            Image(systemName: "globe")
-                .imageScale(.large)
-                .foregroundStyle(.tint)
-            Text("Hello, world!")
-        }
-        .padding()
-    }
-}
-
-#Preview {
-    ContentView()
-}
