+++ WeatherForecastApp/WeatherForecastApp/Enum/ViewState.swift
@@ -1,0 +1,14 @@
+//
+//  ViewState.swift
+//  WeatherForecastApp
+//
+//  Created by Lawrence on 10/30/23.
+//
+
+import Foundation
+
+enum ViewState {
+    case loading
+    case loaded
+    case error
+}
